HIVE-4525 : Support timestamps earlier than 1970 and later than 2038 (Mikhail Bautin via Ashutosh Chauhan)
git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1508537 13f79535-47bb-0310-9956-ffa450edef68

+++ ql/src/java/org/apache/hadoop/hive/ql/udf/UDFToInteger.java
@@ -184,7 +184,7 @@ public class UDFToInteger extends UDF {
     if (i == null) {
       return null;
     } else {
-      intWritable.set(i.getSeconds());
+      intWritable.set((int) i.getSeconds());
       return intWritable;
     }
   }
